+++ lib/ZarrCore/src/Codecs/V3/public_names_v3.jl
@@ -2,6 +2,5 @@
 # these are the extension points for defining and registering new v3 codecs.
 public V3Codec, getCodec, register_codec, codec_parsers, codec_encode,
     codec_decode, is_fixed_size, name
-public BloscCodec, BytesCodec, CRC32cCodec, GzipCodec, ShardingCodec,
-    TransposeCodec, GzipV3Codec, BloscV3Codec, ZstdV3Codec, CRC32cV3Codec,
+public BytesCodec, CRC32cCodec, ShardingCodec, TransposeCodec, CRC32cV3Codec,
     VLenUTF8V3Codec